Added tensor file as input.
The tensor file format is a list of floats separated by spaces:

xDimension yDimension zDimension minValue maxValue { (xDimension*yDimension*zDimension) function values } EoF

The first three numbers are the size of the tensor.  The second two are the maximum and minimum values of all the
numbers in the tensor, then there is a list of the values incrementing x fastest, then y, then z.

$ ./dmc -tensor MyTensorFile